+++ src/ffi_gen_tmp.h
@@ -3,15 +3,13 @@
 
 // copy here the definition you would like to generate.
 // go into "flutter_soloud" dir from the root project dir
+// and run:
 //
-// flutter pub run ffigen --config ffigen_player.yaml
-// to generate [flutter_soloud_bindings_ffi_TMP.dart] or
+// flutter pub run ffigen --config ffigen.yaml
+// to generate [flutter_soloud_FFIGEN.dart]
 //
-// flutter pub run ffigen --config ffigen_capture.yaml
-// to generate [lib/bindings_capture_ffi_TMP.dart]
-//
-// the generated code will be in flutter_soloud_bindings_ffi_TMP.dart
-// copy the generated definition  into flutter_soloud_bindings_ffi.dart
+// the generated code will be placed into flutter_soloud_FFIGEN.dart
+// copy the generated definition into flutter_soloud_bindings_ffi.dart
 
 #include "enums.h"
 
@@ -25,9 +23,8 @@ struct CaptureDevice
 
 //--------------------- copy here the new functions to generate
 
-/// This function can be used to set a sample to play on repeat, 
-/// instead of just playing once
-///
-/// [soundHash]
-/// [enable]
-FFI_PLUGIN_EXPORT void setLooping(unsigned int soundHash, bool enable);
+    /// Dispose all sounds already loaded
+    ///
+    FFI_PLUGIN_EXPORT void disposeAllSound();
+
+